+++ create_dataset/models.py
@@ -11,3 +11,7 @@ class Dataset(models.Model):
     task_id = models.TextField(null=True)
     errors = JSONField(blank=True, null=True)
     warnings = JSONField(blank=True, null=True)
+    # genbank datasets require nucleotidae and aminoacid datasets
+    sister_dataset_id = models.IntegerField(null=True, blank=True)
+    # eg. Phylip datasets require an extra part for the gene definitions
+    charset_block = models.TextField(null=True)